fix: improve process management and clipboard synchronization
The previous method of identifying the recording process using `pgrep arecord` was unreliable as it could terminate unrelated processes. This commit introduces a PID file to specifically track and manage the recording process instance.

Additionally, timeouts have been added to the process termination and clipboard synchronization loops to prevent the script from hanging indefinitely in case of failure.